refactor: rename utils to strings
